UWP transport fix to allow plain socket connection
